refactor modularize documentparser and fix dependencies
