+++ src/Bookmarklet/index.js
@@ -1,6 +1,5 @@
+import documentParser from '../ContentScript/documentParser';
 import { defaultPrefs } from '../Preferences';
-import contentStyle from '../ContentScript/contentStyle.scss';
-import { setReadingMode, addStyles } from '../ContentScript';
 
 const { saccadesInterval, fixationStrength } = { ...defaultPrefs, fixationStrength: 3 };
 
@@ -9,6 +8,4 @@ document.body.setAttribute('fixation-strength', fixationStrength);
 
 document.body.classList.toggle('br-bold');
 
-addStyles(contentStyle.toString());
-
-setReadingMode(document.body.classList.contains('br-bold'));
+documentParser.setReadingMode(document.body.classList.contains('br-bold'));
